Remove bad docstring content
This is confusing and irrelevant -- the position of an axis is
defined by its origin and not by its co-ordinate space.

This class should represent a generic coordinate rather than
trying to define the space too. We can define the space
elsewhere and achieve both greated clarity and reduce repetition.