Add support for generic arduino commands
We take in a command as a list of JSON-encodable entities and
pass that to robotd. We expect to immediately get back a response
message (as well as the usual status blob) which contains either
valid data from the arduino or a description of the failure.

This includes a novel testing approach which:
- creates a valid robotd Board instance
- mocks the underlying serial connection
- creates a robot-api Board instance
- mocks the socket connecting the robot-api Board such that it
  connects to the robotd Board instance

This avoids real sockets and multiprocessing, both of which incur
substantial overhead during tests along with the introduction of
timeouts (which is just nasty).