Simplify robotd connections
This copies over the Connection class from robotd (extending it
with types) and uses it to simplify our connection handling. Our
version has a notable difference that when it receives an empty
message it raises a BrokenPipeError so that this failure is loud.

This commit deliberately strips back the retrying support to a
minimal single retry which only occurs on errors which are actually
fixable by reconnecting. Specifically this means that we now
_don't_ catch:
 - socket.timeout: this would need a very different retry approach
   than we currently use (we should retry using the same socket,
   rather than a new one) and isn't likely to be resolvable given
   that we have a long timeout anyway
 - OSError: the only error of this class we care about (as far as
   I'm aware) is FileNotFoundError, which wouldn't be thrown by
   either the `send` or `receive` methods anyway (only when connecting).