Collect the board lists only when needed
This ensures that they actually contain the lists of boards at
this point, thus ensuring that we properly close off the boards.

Note: while it might be assumed that this could be done upfront,
we reassign the `self.known*boards` attributes when adding boards,
meaning that the upfront collection will only reflect the initial
(empty) lists.