Make _update_boards do all the things
This changes _update_boards so that it modifies the lists of boards
in-place rather than returning them and has it instead return a
`BoardList[T]` of all boards. This allows us to reduce the logic
in the accessor properties about as far as is possible.

Remove the TODO tag now that we've reduced the boards' properties
to their realistic minimum, even though we've not quite achieved
what it suggests.