tests: refactor temporary directory creation into a class mixin
As suggested in https://github.com/sourcebots/robot-api/pull/64#discussion_r165854809

This also switches to using a `tempfile.TemporaryDirectory` instead
of calling `tempfile.mkdtemp` directly, which is safer since the
former ensures the directory is cleaned up even if you forget to
explicitly tell it do (a warning is displayed if this does occur).

Additionally, `unittest.TestCase.addCleanup` is used to schedule
cleanup operations at the time of resource creation, removing the
need for cleanup logic to be duplicated in each test case's
`tearDown` method.